Bring updated dropdown to Combobox
Width is still broken for reasons
